refactor Key definition
