+++ src/codes_handle/mod.rs
@@ -36,29 +36,6 @@ pub struct CodesHandle<SOURCE: Debug + SpecialDrop> {
     unsafe_message: KeyedMessage,
 }
 
-///Structure representing a single key from the `KeyedMessage`.
-#[derive(Clone, Debug, PartialEq)]
-pub struct Key {
-    pub name: String,
-    pub value: KeyType,
-}
-
-///Enum to represent and contain all possible types of keys inside `KeyedMessage`.
-///
-///Messages inside GRIB files can contain arbitrary keys set by the file author.
-///The type of a given key is only known at runtime (after being checked).
-///There are several possible types of keys, which are represented by this enum
-///and each variant contains the respective data type.
-#[derive(Clone, Debug, PartialEq)]
-pub enum KeyType {
-    Float(f64),
-    Int(i64),
-    FloatArray(Vec<f64>),
-    IntArray(Vec<i64>),
-    Str(String),
-    Bytes(Vec<u8>),
-}
-
 #[derive(Debug)]
 enum DataContainer {
     FileBytes(Bytes),
